Move and deprecate InMemoryMetricExporter
