+++ sdk/metrics-testing/src/main/java/io/opentelemetry/sdk/testing/exporter/package-info.java
@@ -1,0 +1,9 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+@ParametersAreNonnullByDefault
+package io.opentelemetry.sdk.testing.exporter;
+
+import javax.annotation.ParametersAreNonnullByDefault;
